Change torrent download. Refactor

+++ src/Aria4net/Common/ProcessStarter.cs
@@ -1,17 +1,20 @@
 using System;
 using System.Diagnostics;
+using NLog;
 
 namespace Aria4net.Common
 {
     public abstract class ProcessStarter : IProcessStarter
     {
         private readonly IFileFinder _fileFinder;
+        protected readonly Logger Logger;
         private readonly Process _process;
 
 
-        protected ProcessStarter(IFileFinder fileFinder)
+        protected ProcessStarter(IFileFinder fileFinder, Logger logger)
         {
             _fileFinder = fileFinder;
+            Logger = logger;
             _process = new Process();
         }
 
@@ -24,6 +27,7 @@ namespace Aria4net.Common
 
         public void Run()
         {
+            
             _process.StartInfo.FileName = _fileFinder.Find();
             _process.StartInfo.Arguments = GetArguments();
             _process.EnableRaisingEvents = true;
@@ -32,13 +36,17 @@ namespace Aria4net.Common
             
             ConfigureProcess(_process);
 
+            Logger.Info("Iniciando processo {0}", _process.StartInfo.FileName);
             _process.Start();
         }
 
         public void Exit()
         {
-            _process.Kill();
+            Logger.Info("Encerrando processo {0}", _process.StartInfo.FileName);
+            if(IsRunning()) _process.Kill();
         }
+
+        public abstract bool IsRunning();
 
         protected abstract void ProcessOnExited(Process sender, EventArgs eventArgs);
